src/test/java/de/isys/demo/IndexControllerTest.java
package de.isys.demo;

import org.junit.Test;

import java.util.Map;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.Matchers.anyString;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

public class IndexControllerTest {

	@Test
	public void testGreeting() {
		final GreeterService greeterMock = mock(GreeterService.class);
		when(greeterMock.transformation(anyString())).thenReturn("mock implementation");
		IndexController indexController = new IndexController(
				greeterMock
		);

		final Map<String, Object> model = indexController.greet("spring boot").getModel();
		final Map<String, Object> model2 = indexController.greet("something else").getModel();

		final String nameToGreet = (String) model.get("name_to_greet");
		assertThat(nameToGreet).isEqualTo("mock implementation");
		assertThat(nameToGreet).isNotNull();

		final String nameToGreet2 = (String) model2.get("name_to_greet");
		assertThat(nameToGreet2).isEqualTo("mock implementation");
		assertThat(nameToGreet2).isNotNull();

	}

}
